#943 - refactoring Call components into REST services
